fix: Backend bug workaround.
The backend has a bug where it emits search results to all sockets and
does not include any sort of identifier that can be used to associate
the results with a particular POST request to /startSearch. This means
that if you search for something with lots of results and then quickly
go back and search for something else, we will ultimately get the
results from the first search intermixed with our new search. This
happens despite destroying the socket and related observables because
once we setup new observables and reconnect to the socket server, the
backend will continue to send results from the previous search to all
connected clients, including the brand new socket we just created for
the second search.

This backend demo app is simply incomplete and this app would not work
for more than one concurrent user at a time as it is currently written.
The change in this commit simply disables the "go back" button until we
receive a 'complete' event over the current socket. This is a bandaid
and doesn't prevent the user from manually going back to the root view
and initiating a new search before the other one is done. I figured this
is good enough for the demo considering that this problem would be
immediately obvious if more than one person was trying to use the app
simultaneously. The proper fix would be to modify the backend so it
generates some kind of unique requestId in /startSearch, returns that ID
in the response to the POST call, and then includes the ID in the
emitted results so the client could scrub unrelated results. Or better
yet, create dynamic socket.io topics and only emit the results to the
relevant clients instead of broadcasting results to all connected
clients.